// src/Garage-TraeUI/backend/src/types.ts
export interface GarageMetrics {
  cluster: {
    nodes_total: number;
    nodes_up: number;
    layout_version: number;
  };
  storage: {
    total_capacity: number;
    used_capacity: number;
    available_capacity: number;
  };
  buckets: {
    total_count: number;
  };
  objects: {
    total_count: number;
    total_size: number;
  };
}